chore: added description to endpoints, some queries optimized, some bugs fixed in analytics and attendance groups

+++ adminpage/api/serializers/attendance.py
@@ -83,7 +83,7 @@ class HoursInfoFullSerializer(serializers.Serializer):
 
 
 class BetterThanInfoSerializer(serializers.Serializer):
-    better_than = serializers.FloatField()
+    better_than = serializers.IntegerField()
 
 
 class AttendanceSerializer(serializers.Serializer):
